валидатор

+++ internal/pkg/models/session.go
@@ -2,5 +2,5 @@ package models
 
 type Session struct {
 	Id     string
-	UserId int64
+	UserId int
 }
